requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
pydantic==2.11.3
google-genai==1.8.0
python-dotenv==1.1.0
mcp==1.6.0
pillow==11.1.0
pywinauto==0.6.9
pyautogui==0.9.53
pypi-json==0.4.0
rich==13.5.2
requests>=2.31.0
markitdown[all]>=0.1.1
faiss-cpu>=1.10.0
tqdm>=4.67.1